+++ index.js
@@ -1,7 +1,7 @@
-import {commandValidator, getUserName, goodBye, helloUser} from "./utils.js";
+import {commandValidator, getUserName, goodBye, helloUser} from "./utils/utils.js";
 import * as readline from "node:readline";
-import {EXIT_COMMAND} from "./constants.js";
-import {ActionController} from "./action.controller.js";
+import {EXIT_COMMAND} from "./constants/constants.js";
+import {ActionController} from "./actions-controller/action.controller.js";
 
 const {stdin: input, stdout: output} = process;
 const user = getUserName();
